+++ source/ftrack_connect_3dsmax/ui/info.py
@@ -3,8 +3,8 @@
 
 from QtExt import QtCore, QtGui, QtWidgets
 
-from ftrack_connect.ui.widget.info import FtrackInfoDialog
-from ftrack_connect.ui.theme import applyTheme
+from ftrack_connector_legacy.ui.widget.info import FtrackInfoDialog
+from ftrack_connector_legacy.ui.theme import applyTheme
 
 
 class FtrackMaxInfoDialog(FtrackInfoDialog):